pyproject.toml
[tool.poetry]
name = "my-test-task"
version = "0.1.0"
description = ""
authors = ["Oleg Lomtev <lomtev.oa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
mysql-connector-python = "^9.1.0"


[tool.poetry.group.lint.dependencies]
mypy = "^1.13.0"
flake8 = "^7.1.1"
black = "^24.10.0"
isort = "^5.13.2"


[tool.poetry.group.DB.dependencies]
pandas = "^2.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = 'venv'

[tool.black]
line-length = 119
exclude = '.git'

[tool.isort]
line_length = 119